fixed integer overflow operation in iwavsk
despite `iwavsk` function being declared as integer(8), the computations is contained were still using 4 bytes integer instead of 8 byte integers. This caused `iwavsk` result to be negative (`lrairy` was about 39mb (39513784) and with `irecr` = 59,  seek position exceeded the maximum positive value of a 4-byte integer, resulting in a negative value -2141459559)

Now the computations are performed using 8-byte integer arithmetics

fixes issue #110

+++ lib/hiwav.F90
@@ -14,7 +14,7 @@ module mod_wave  ! mod_wav is a modern replacement of common block cowave
   contains
 function get_wfu_rec1_length(nchwfu)
   integer, intent(in) :: nchwfu  ! number of channels in wfu file
-  integer :: get_wfu_rec1_length
+  integer(8) :: get_wfu_rec1_length
   !     The three variables below are used to determine the (machine
   !     dependent) size of the built-in types
   character :: char_t
